Complete desktop view for /about pages
