+++ package.json
@@ -17,30 +17,30 @@
   },
   "dependencies": {
     "@changesets/cli": "^2.26.2",
-    "@melt-ui/svelte": "^0.34.0",
-    "@sanity/client": "^6.4.3",
+    "@melt-ui/svelte": "^0.34.2",
+    "@sanity/client": "^6.4.4",
     "@sanity/groq-store": "^4.0.3",
     "@sanity/icons": "^2.4.1",
     "@sanity/image-url": "^1.0.2",
     "@sanity/ui": "^1.7.4",
-    "@sanity/vision": "^3.14.5",
+    "@sanity/vision": "^3.15.0",
     "@sanity/webhook": "^2.0.0",
     "@vercel/analytics": "^1.0.1",
-    "groq": "^3.14.5",
+    "groq": "^3.15.0",
     "lucide-react": "^0.265.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-is": "^18.2.0",
-    "sanity": "^3.14.5",
+    "sanity": "^3.15.0",
     "styled-components": "^5.3.11"
   },
   "devDependencies": {
-    "@melt-ui/pp": "^0.1.1",
+    "@melt-ui/pp": "^0.1.2",
     "@playwright/test": "^1.36.2",
     "@portabletext/svelte": "^2.0.0",
     "@sveltejs/adapter-vercel": "^3.0.3",
     "@sveltejs/kit": "^1.22.4",
-    "@types/react": "^18.2.18",
+    "@types/react": "^18.2.19",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.3.0",
     "@typescript-eslint/parser": "^6.3.0",
