+++ src/lib/components/portable-text/index.ts
@@ -1,0 +1,2 @@
+export { default as UnorderedList } from './UnorderedList.svelte';
+export { default as ListItem } from './ListItem.svelte';
