+++ src/lib/config/sanity/schemas/index.ts
@@ -3,7 +3,7 @@ export { Company } from './company';
 export { Education } from './education';
 export { Experience } from './experience';
 export { Project } from './project';
-// export { Skill } from './skill';
+export { School } from './school';
 export { Skills } from './skills';
 export { User } from './user';
