+++ src/routes/(app)/about/+page.server.ts
@@ -1,0 +1,6 @@
+import type { PageServerLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
+
+export const load = (async () => {
+  throw redirect(307, '/about/general');
+}) satisfies PageServerLoad;
