+++ src/routes/(app)/about/general/+page.server.ts
