--- src/routes/(app)/resume/projects/+page.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-  import type { PageData } from './$types';
-
-  export let data: PageData;
-</script>
